requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
fastapi==0.133.1
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
multidict==6.7.1
mysqlclient==2.2.8
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
qrcode==8.2
razorpay==2.0.0
requests==2.32.5
SQLAlchemy==2.0.47
sqlparse==0.5.5
starlette==0.52.1
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
yarl==1.22.0
